pyproject.toml
[tool.poetry]
name = "vulnhuntr"
version = "0.1.0"
description = ""
authors = ["Dan <dan@protectai.com>", "Marcello <marcello@protectai.com>"]
readme = "README.md"

[tool.poetry.scripts]
vulnhuntr = 'vulnhuntr.__main__:run'

[tool.poetry.dependencies]
python = "^3.10"
rich = "^13.7.1"
pydantic-xml = "^2.11.0"
pydantic = "^2.8.0"
anthropic = "^0.30.1"
structlog = "^24.2.0"
jedi = "0.18.0"
parso = "0.8.0"
openai = "^1.51.2"
google-generativeai = "^0.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
